+++ atm-service/src/main/java/com/github/jukkarol/dto/withdrawalDto/request/MakeWithdrawalRequest.java
@@ -5,10 +5,12 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record MakeWithdrawalRequest(
         @NotNull
         @Positive
-        Integer amount,
+        BigDecimal amount,
 
         @Size(min = 10, max = 10)
         @NotEmpty
